refac: ajustando atributos editáveis, fluxo de pedido, tempo de produção no MES

+++ app_smart_4.0/src/main/java/com/tecdes/smart/app_smart_40/model/clp/EstoqueCLP.java
@@ -51,7 +51,9 @@ public class EstoqueCLP extends EstacaoCLP {
     private boolean iniciarGuardarEst;
     private int posicaoGuardarEst;
 
-    private byte[] posicoesOcupadas;
+    // int[] (não byte[]): serializa sempre como vetor de números [0,1,0,2,...] (cada posição = cor do
+    // magazine). Com byte[] o Jackson emitiria Base64 (string), que chegava como "caracteres" no front.
+    private int[] posicoesOcupadas;
 
     private boolean pedirPosicaoEst;
     private int posicaoEstoque;
